refactor: the view part of nodeDetails -> NodeDetails
