--- src/galaxy/nodeDetails/templates/all.js
@@ -1,9 +1,0 @@
-/**
- * TODO: I need something better than this. Manually changing template is
- * very inconvenient :(. Should it be defined with metadata?
- */
-import defaultTemplate from './default.jsx';
-
-export default {
-  default: defaultTemplate
-}
