+++ src/styles/nodeDetails.less
@@ -2,7 +2,7 @@
 @line-height: 1.4;
 @lines-to-show: 2;
 
-.node-details {
+.basic-node-info {
   width: 395px;
   left: 16px;
   bottom: 12px;
